Temporairly commented-out extraneous modules
The POSIX and Linux platform controllers are required for Windows OSes, and
to get to the heart of the matter quickly they have been removed while the
windows controller is developed.  The removed modules will have to be added
back into the codepath under conditional or exceptional paths once a
functioning windows controller is ready.